requirements.txt
streamlit
pandas
pypdf
python-docx
fpdf
langchain-google-genai
python-dotenv
